cleanup and improve the inspector tests
